Resolve soft pinned version to nearest version if it does not exist.
Currently, even if a soft pinned version, e.g. "3.0", does not
exist aether returns that soft pinned version. Instead, it should
provide another valid version. This change solves this issue by
transforming the soft pin into a version range, and making the
aether request using that version range.

For example, instead of requesting for "3.0", we transform it to "[3.0,)"
providing us all versions between 3.0 and the latest version.
If the version exists, we simply take it. If it doesn't, then
we select the next closest valid version.